Moved threshold tests together
